moved getdefaultencoding from text to py3compat
