+++ IPython/utils/_process_win32.py
@@ -94,7 +94,7 @@ def _find_cmd(cmd):
 
 def _system_body(p):
     """Callback for _system."""
-    enc = text.getdefaultencoding()
+    enc = py3compat.getdefaultencoding()
     for line in read_no_interrupt(p.stdout).splitlines():
         line = line.decode(enc, 'replace')
         print(line, file=sys.stdout)
