+++ IPython/utils/io.py
@@ -14,6 +14,7 @@ from __future__ import print_function
 #-----------------------------------------------------------------------------
 # Imports
 #-----------------------------------------------------------------------------
+import os
 import sys
 import tempfile
