+++ IPython/utils/jsonutil.py
@@ -135,7 +135,7 @@ def json_clean(obj):
         return obj
     
     if isinstance(obj, bytes):
-        return obj.decode(text.getdefaultencoding(), 'replace')
+        return obj.decode(py3compat.getdefaultencoding(), 'replace')
     
     if isinstance(obj, container_to_list) or (
         hasattr(obj, '__iter__') and hasattr(obj, next_attr_name)):
