+++ IPython/utils/text.py
@@ -16,7 +16,6 @@ Utilities for working with strings and text.
 
 import __main__
 
-import locale
 import os
 import re
 import shutil
@@ -33,29 +32,6 @@ from IPython.utils.data import flatten
 #-----------------------------------------------------------------------------
 # Code
 #-----------------------------------------------------------------------------
-
-# Less conservative replacement for sys.getdefaultencoding, that will try
-# to match the environment.
-# Defined here as central function, so if we find better choices, we
-# won't need to make changes all over IPython.
-def getdefaultencoding():
-    """Return IPython's guess for the default encoding for bytes as text.
-
-    Asks for stdin.encoding first, to match the calling Terminal, but that
-    is often None for subprocesses.  Fall back on locale.getpreferredencoding()
-    which should be a sensible platform default (that respects LANG environment),
-    and finally to sys.getdefaultencoding() which is the most conservative option,
-    and usually ASCII.
-    """
-    enc = py3compat.get_stream_enc(sys.stdin)
-    if not enc or enc=='ascii':
-        try:
-            # There are reports of getpreferredencoding raising errors
-            # in some cases, which may well be fixed, but let's be conservative here.
-            enc = locale.getpreferredencoding()
-        except Exception:
-            pass
-    return enc or sys.getdefaultencoding()
 
 def unquote_ends(istr):
     """Remove a single pair of quotes from the endpoints of a string."""
